Fix duplicated interfaces in mtsj
