--- java/mtsj/api/src/main/java/com/devonfw/application/mtsj/general/common/api/Usermanagement.java
@@ -1,15 +1,0 @@
-package com.devonfw.application.mtsj.general.common.api;
-
-/**
- * Interface to get a user from its login.
- *
- */
-public interface Usermanagement {
-
-  /**
-   * @param login The login of the requested user.
-   * @return The {@link UserProfile} with the given <code>login</code> or {@code null} if no such object exists.
-   */
-  UserProfile findUserProfileByLogin(String login);
-
-}
