+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/general/common/impl/security/BaseUserDetailsService.java
@@ -20,7 +20,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
 import com.devonfw.application.mtsj.general.common.api.UserProfile;
-import com.devonfw.application.mtsj.general.common.api.Usermanagement;
+import com.devonfw.application.mtsj.usermanagement.logic.api.Usermanagement;
 import com.devonfw.application.mtsj.general.common.base.BaseUserDetails;
 import com.devonfw.application.mtsj.usermanagement.dataaccess.api.UserEntity;
 import com.devonfw.application.mtsj.usermanagement.dataaccess.api.repo.UserRepository;
