--- java/mtsj/core/src/main/java/com/devonfw/application/mtsj/general/logic/impl/UsermanagementDummyImpl.java
@@ -1,32 +1,0 @@
-package com.devonfw.application.mtsj.general.logic.impl;
-
-import javax.inject.Named;
-
-import org.springframework.stereotype.Component;
-
-import com.devonfw.application.mtsj.general.common.api.UserProfile;
-import com.devonfw.application.mtsj.general.common.api.Usermanagement;
-import com.devonfw.application.mtsj.general.common.api.datatype.Role;
-import com.devonfw.application.mtsj.general.common.api.to.UserDetailsClientTo;
-import com.devonfw.application.mtsj.general.common.base.AbstractBeanMapperSupport;
-
-/**
- * Implementation of {@link Usermanagement}.
- */
-@Named
-@Component
-public class UsermanagementDummyImpl extends AbstractBeanMapperSupport implements Usermanagement {
-
-  @Override
-  public UserProfile findUserProfileByLogin(String login) {
-
-    // this is only a dummy - please replace with a real implementation
-    UserDetailsClientTo profile = new UserDetailsClientTo();
-    profile.setName(login);
-    profile.setFirstName("Peter");
-    profile.setLastName(login);
-    profile.setRole(Role.WAITER);
-    return profile;
-  }
-
-}
